Inline initiateLinting function for bundling perf
